stroytechmontazh/requirements.txt
apispec==6.3.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-appconf==1.0.5
django-compressor==4.4
django-debug-toolbar==4.2.0
django-sass-processor==1.2.2
django-serializer==1.2.1
djangorestframework==3.14.0
gunicorn==21.2.0
httptools==0.6.1
idna==3.4
libsass==0.22.0
marshmallow==3.20.1
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
sqlparse==0.4.4
Unit==0.2.2
urllib3==2.0.7
